+++ cpp/daal/src/algorithms/linear_model/linear_model_train_normeq_finalize_impl.i
@@ -26,7 +26,7 @@
 
 #include "src/algorithms/service_kernel_math.h"
 #include "src/externals/service_lapack.h"
-#include "services/service_profiler.h"
+#include "services/internal/service_profiler.h"
 
 namespace daal
 {
