+++ cpp/daal/src/algorithms/linear_model/linear_model_train_normeq_update_impl.i
@@ -27,7 +27,7 @@
 #include "src/externals/service_blas.h"
 #include "src/algorithms/service_error_handling.h"
 #include "src/threading/threading.h"
-#include "services/service_profiler.h"
+#include "services/internal/service_profiler.h"
 
 namespace daal
 {
