+++ cpp/daal/src/algorithms/pca/pca_dense_correlation_batch_impl.i
@@ -30,7 +30,7 @@
 #include "src/algorithms/service_error_handling.h"
 #include "src/threading/threading.h"
 
-#include "services/service_profiler.h"
+#include "services/internal/service_profiler.h"
 
 namespace daal
 {
